Rename TestCase to MonologTestCase
