+++ src/Monolog/Test/TestCase.php
@@ -24,51 +24,8 @@ use ReflectionProperty;
  *
  * @author Jordi Boggiano <j.boggiano@seld.be>
  *
- * @internal feel free to reuse this to test your own handlers, this is marked internal to avoid issues with PHPStorm https://github.com/Seldaek/monolog/issues/1677
+ * @deprecated use MonologTestCase instead.
  */
-class TestCase extends \PHPUnit\Framework\TestCase
+class TestCase extends MonologTestCase
 {
-    /**
-     * @param array<mixed> $context
-     * @param array<mixed> $extra
-     *
-     * @phpstan-param value-of<Level::VALUES>|value-of<Level::NAMES>|Level|LogLevel::* $level
-     */
-    protected function getRecord(int|string|Level $level = Level::Warning, string|\Stringable $message = 'test', array $context = [], string $channel = 'test', \DateTimeImmutable $datetime = new JsonSerializableDateTimeImmutable(true), array $extra = []): LogRecord
-    {
-        return new LogRecord(
-            message: (string) $message,
-            context: $context,
-            level: Logger::toMonologLevel($level),
-            channel: $channel,
-            datetime: $datetime,
-            extra: $extra,
-        );
-    }
-
-    /**
-     * @phpstan-return list<LogRecord>
-     */
-    protected function getMultipleRecords(): array
-    {
-        return [
-            $this->getRecord(Level::Debug, 'debug message 1'),
-            $this->getRecord(Level::Debug, 'debug message 2'),
-            $this->getRecord(Level::Info, 'information'),
-            $this->getRecord(Level::Warning, 'warning'),
-            $this->getRecord(Level::Error, 'error'),
-        ];
-    }
-
-    protected function getIdentityFormatter(): FormatterInterface
-    {
-        $formatter = $this->createMock(FormatterInterface::class);
-        $formatter->expects(self::any())
-            ->method('format')
-            ->willReturnCallback(function ($record) {
-                return $record->message;
-            });
-
-        return $formatter;
-    }
 }
